Drop boilerplate by relying on blanket impls for `link_kind` hostfunc.

+++ pangalactic_derivelib/src/bindings.rs
@@ -1,4 +1,4 @@
 #[link(wasm_import_module = "pangalactic_bindings")]
 extern "C" {
-    pub(crate) fn link_kind(handle: i64) -> i32;
+    pub(crate) fn link_kind(handle: i64) -> i64;
 }
